# contracts/openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Reference/ability-scores": {
      "get": {
        "tags": [
          "Reference"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AbilityScoreDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AbilityScoreDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AbilityScoreDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reference"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAbilityScoreRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAbilityScoreRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAbilityScoreRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AbilityScoreDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbilityScoreDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbilityScoreDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Reference/races": {
      "get": {
        "tags": [
          "Reference"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RaceDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RaceDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RaceDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reference"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRaceRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRaceRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRaceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RaceDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RaceDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RaceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reference/races/{id}": {
      "get": {
        "tags": [
          "Reference"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RaceDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RaceDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RaceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reference/classes": {
      "get": {
        "tags": [
          "Reference"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reference"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClassDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reference/classes/{id}": {
      "get": {
        "tags": [
          "Reference"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClassDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reference/items": {
      "get": {
        "tags": [
          "Reference"
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ItemType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemSummaryDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemSummaryDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemSummaryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reference"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reference/items/{id}": {
      "get": {
        "tags": [
          "Reference"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reference/spells": {
      "get": {
        "tags": [
          "Reference"
        ],
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "schoolId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "classId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpellSummaryDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpellSummaryDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpellSummaryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reference"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpellRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpellRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpellRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpellDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reference/spells/{id}": {
      "get": {
        "tags": [
          "Reference"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpellDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reference/monsters": {
      "get": {
        "tags": [
          "Reference"
        ],
        "parameters": [
          {
            "name": "maxCr",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "creatureTypeId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonsterSummaryDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonsterSummaryDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonsterSummaryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reference"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMonsterRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMonsterRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMonsterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MonsterDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonsterDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonsterDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reference/monsters/{id}": {
      "get": {
        "tags": [
          "Reference"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MonsterDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonsterDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonsterDetailDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AbilityBonusDto": {
        "type": "object",
        "properties": {
          "abilityScoreName": {
            "type": "string",
            "nullable": true
          },
          "abbreviation": {
            "type": "string",
            "nullable": true
          },
          "bonus": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "AbilityScoreDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "abbreviation": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isCustom": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "CasterType": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ClassDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "hitDie": {
            "type": "integer",
            "format": "int32"
          },
          "spellcastingAbility": {
            "type": "string",
            "nullable": true
          },
          "casterType": {
            "$ref": "#/components/schemas/CasterType"
          },
          "isCustom": {
            "type": "boolean"
          },
          "savingThrows": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "features": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClassFeatureDto"
            },
            "nullable": true
          },
          "subclasses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubclassDto"
            },
            "nullable": true
          },
          "spellSlotProgression": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpellSlotRowDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateAbilityScoreRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "abbreviation": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateClassRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "hitDie": {
            "type": "integer",
            "format": "int32"
          },
          "spellcastingAbilityScoreId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "casterType": {
            "$ref": "#/components/schemas/CasterType"
          },
          "savingThrowAbilityIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateItemRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "weightLbs": {
            "type": "number",
            "format": "double"
          },
          "costCp": {
            "type": "integer",
            "format": "int32"
          },
          "itemType": {
            "$ref": "#/components/schemas/ItemType"
          },
          "weapon": {
            "$ref": "#/components/schemas/WeaponInput"
          },
          "armor": {
            "$ref": "#/components/schemas/ArmorInput"
          }
        },
        "additionalProperties": false
      },
...
      "CreateRaceRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "speed": {
            "type": "integer",
            "format": "int32"
          },
          "sizeId": {
            "type": "integer",
            "format": "int32"
          },
          "abilityBonuses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AbilityBonusInput"
            },
            "nullable": true
          },
          "traits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TraitInput"
            },
            "nullable": true
          },
          "languageIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ItemDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "itemType": {
            "$ref": "#/components/schemas/ItemType"
          },
          "weightLbs": {
            "type": "number",
            "format": "double"
          },
          "costCp": {
            "type": "integer",
            "format": "int32"
          },
          "isCustom": {
            "type": "boolean"
          },
          "weapon": {
            "$ref": "#/components/schemas/WeaponDetailDto"
          },
          "armor": {
            "$ref": "#/components/schemas/ArmorDetailDto"
          },
          "magicItem": {
            "$ref": "#/components/schemas/MagicItemDetailDto"
          }
        },
        "additionalProperties": false
      },
...
      "ItemSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "itemType": {
            "$ref": "#/components/schemas/ItemType"
          },
          "weightLbs": {
            "type": "number",
            "format": "double"
          },
          "costCp": {
            "type": "integer",
            "format": "int32"
          },
          "isCustom": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ItemType": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "RaceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "speed": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "string",
            "nullable": true
          },
          "isCustom": {
            "type": "boolean"
          },
          "abilityBonuses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AbilityBonusDto"
            },
            "nullable": true
          },
          "traits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TraitDto"
            },
            "nullable": true
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "subraces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubraceDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SpellSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "school": {
            "type": "string",
            "nullable": true
          },
          "castingTime": {
            "type": "string",
            "nullable": true
          },
          "isConcentration": {
            "type": "boolean"
          },
          "isRitual": {
            "type": "boolean"
          },
          "isCustom": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}